+++ templates/admin-panel/template/server/{{shortName}}/admin_members_test.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"{{ shortName }}/models"
 	"path"
-	"test_helpers"
+	"{{ shortName }}/test_helpers"
 	"testing"
 )