+++ templates/admin-panel/template/server/{{shortName}}/admin_server_info_test.go
@@ -7,11 +7,11 @@ import (
 	"net/http"
 	"{{ shortName }}/models"
 	"path"
-	"test_helpers"
+	"{{ shortName }}/test_helpers"
 	"testing"
 )
 
-func TestGetSuMemberListBundle(t *testing.T) {
+func TestGetServerInfo(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
@@ -26,27 +26,29 @@ func TestGetSuMemberListBundle(t *testing.T) {
 	loginResponse := test_helpers.ExecuteRequest(n, loginReq)
 	test_helpers.Equals(t, http.StatusOK, loginResponse.Code)
 
-	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
+	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
 	req.Header.Set("Cookie", test_helpers.MergeCookies(loginReq, loginResponse))
 	req.Header.Set("X-CSRF-Token", loginResponse.Header().Get("X-CSRF-Token"))
 	response := test_helpers.ExecuteRequest(n, req)
 	test_helpers.Equals(t, http.StatusOK, response.Code)
-	test_helpers.Equals(t, true, len(response.Body.Bytes()) > 0)
+	_, err := test_helpers.UnmarshalJSON(response.Body.Bytes())
+	test_helpers.Ok(t, err)
 }
 
-func TestGetSuMemberListBundleNonAutorized(t *testing.T) {
+func TestGetServerInfoNonAutorized(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
 	test_helpers.ClearDB(models.Db)
-	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
+	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
 	test_helpers.SetCSRFToken(n, req)
 	response := test_helpers.ExecuteRequest(n, req)
-	test_helpers.Equals(t, http.StatusUnauthorized, response.Code)
-	test_helpers.Equals(t, "Unauthorized\n", string(response.Body.Bytes()))
+	test_helpers.Equals(t, http.StatusOK, response.Code)
+	body, _ := test_helpers.UnmarshalJSON(response.Body.Bytes())
+	test_helpers.Equals(t, map[string]interface{}{"Status": "Expired session or cookie", "Messages": []interface{}{"Session Expired.  Log out and log back in."}}, body)
 }
 
-func TestGetSuMemberListBundleForNonSuperuser(t *testing.T) {
+func TestGetServerInfoForNonSuperuser(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
@@ -61,10 +63,14 @@ func TestGetSuMemberListBundleForNonSuperuser(t *testing.T) {
 	loginResponse := test_helpers.ExecuteRequest(n, loginReq)
 	test_helpers.Equals(t, http.StatusOK, loginResponse.Code)
 
-	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
+	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
 	req.Header.Set("Cookie", test_helpers.MergeCookies(loginReq, loginResponse))
 	req.Header.Set("X-CSRF-Token", loginResponse.Header().Get("X-CSRF-Token"))
 	response := test_helpers.ExecuteRequest(n, req)
 	test_helpers.Equals(t, http.StatusForbidden, response.Code)
-	test_helpers.Equals(t, "Forbidden\n", string(response.Body.Bytes()))
+	body, _ := test_helpers.UnmarshalJSON(response.Body.Bytes())
+	test_helpers.Equals(t, map[string]interface{}{
+		"Messages": []interface{}{"Only superuser can get server info"},
+		"Status":   "Only superuser can get server info",
+	}, body)
 }