+++ templates/admin-panel/template/server/{{shortName}}/assets_test.go
@@ -7,11 +7,11 @@ import (
 	"net/http"
 	"{{ shortName }}/models"
 	"path"
-	"test_helpers"
+	"{{ shortName }}/test_helpers"
 	"testing"
 )
 
-func TestGetServerInfo(t *testing.T) {
+func TestGetSuMemberListBundle(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
@@ -26,29 +26,27 @@ func TestGetServerInfo(t *testing.T) {
 	loginResponse := test_helpers.ExecuteRequest(n, loginReq)
 	test_helpers.Equals(t, http.StatusOK, loginResponse.Code)
 
-	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
+	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
 	req.Header.Set("Cookie", test_helpers.MergeCookies(loginReq, loginResponse))
 	req.Header.Set("X-CSRF-Token", loginResponse.Header().Get("X-CSRF-Token"))
 	response := test_helpers.ExecuteRequest(n, req)
 	test_helpers.Equals(t, http.StatusOK, response.Code)
-	_, err := test_helpers.UnmarshalJSON(response.Body.Bytes())
-	test_helpers.Ok(t, err)
+	test_helpers.Equals(t, true, len(response.Body.Bytes()) > 0)
 }
 
-func TestGetServerInfoNonAutorized(t *testing.T) {
+func TestGetSuMemberListBundleNonAutorized(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
 	test_helpers.ClearDB(models.Db)
-	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
+	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
 	test_helpers.SetCSRFToken(n, req)
 	response := test_helpers.ExecuteRequest(n, req)
-	test_helpers.Equals(t, http.StatusOK, response.Code)
-	body, _ := test_helpers.UnmarshalJSON(response.Body.Bytes())
-	test_helpers.Equals(t, map[string]interface{}{"Status": "Expired session or cookie", "Messages": []interface{}{"Session Expired.  Log out and log back in."}}, body)
+	test_helpers.Equals(t, http.StatusUnauthorized, response.Code)
+	test_helpers.Equals(t, "Unauthorized\n", string(response.Body.Bytes()))
 }
 
-func TestGetServerInfoForNonSuperuser(t *testing.T) {
+func TestGetSuMemberListBundleForNonSuperuser(t *testing.T) {
 	test_helpers.DisableLogger()
 	n := negroni.New()
 	startApplication(path.Join("testdata", "config.cfg"), n)
@@ -63,14 +61,10 @@ func TestGetServerInfoForNonSuperuser(t *testing.T) {
 	loginResponse := test_helpers.ExecuteRequest(n, loginReq)
 	test_helpers.Equals(t, http.StatusOK, loginResponse.Code)
 
-	req, _ := http.NewRequest("GET", "/api/admin/server-info", nil)
+	req, _ := http.NewRequest("GET", "/dist/su_members_list.bundle.js", nil)
 	req.Header.Set("Cookie", test_helpers.MergeCookies(loginReq, loginResponse))
 	req.Header.Set("X-CSRF-Token", loginResponse.Header().Get("X-CSRF-Token"))
 	response := test_helpers.ExecuteRequest(n, req)
 	test_helpers.Equals(t, http.StatusForbidden, response.Code)
-	body, _ := test_helpers.UnmarshalJSON(response.Body.Bytes())
-	test_helpers.Equals(t, map[string]interface{}{
-		"Messages": []interface{}{"Only superuser can get server info"},
-		"Status":   "Only superuser can get server info",
-	}, body)
+	test_helpers.Equals(t, "Forbidden\n", string(response.Body.Bytes()))
 }