+++ templates/admin-panel/template/server/{{shortName}}/serverinfo/handlers.go
@@ -1,7 +1,7 @@
 package serverinfo
 
 import (
-	"assets"
+	"{{ shortName }}/assets"
 	"encoding/json"
 	"net/http"
 	"os"