+++ templates/admin-panel/template/server/{{shortName}}/sessions_test.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"{{ shortName }}/models"
 	"path"
-	"test_helpers"
+	"{{ shortName }}/test_helpers"
 	"testing"
 	"github.com/urfave/negroni"
 	"encoding/json"