+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## master
+## 4.2.0
 * Add page about server information in admin-panel template [#72](https://github.com/sham-ui/sham-ui-cli/issues/72)
 * Fix a tests in admin-panel golang tests
 * Move ErrorPanel component from panel to page in members page in admin-panel template [#73](https://github.com/sham-ui/sham-ui-cli/issues/73)